Fix version stamp script for Py3 compatibility.
Fix version stamp script so that it can be used in Python 3;
specifically, ensure that we decode the bytes taken from popen as a
UTF-8 string, and that we use the print function.

Technically, this is slightly broken on systems that run the script in
encodings other than UTF-8, as .decode() assumes a default encoding of
UTF-8. It'll be too much work for too little pay-off to make this script
work on other encodings.