Allow optional keyword arguments in get().
In Python 3, configparser.get() is more strict and conforms better to
the documented interface for .get, in that `raw` and `vars` are actual
optional paramters. In Python 2, .get() in our parent class doesn't
declare those parameters.

To support both, we simply declare optional keyword arguments wholesale,
and pass those to our superclass when necessary.